Font autogeneration and fixes for modules

+++ TGTools/TGTools/private/Actor.cpp
@@ -1,5 +1,6 @@
 #include "../public/Actor.hpp"
 #include "../public/json.hpp"
+#include "../public/Map.hpp"
 
 namespace tgt::Actor {
 
@@ -31,23 +32,35 @@ namespace tgt::Actor {
 
 	const Result remove(const char* name) {
 		STRING_CHECKS_C(name);
+		return remove(std::string(name));
+	}
 
+	const Result remove(const std::string& name) {
+		STRING_CHECKS(name);
 		auto actor = Util::getResource(ACTOR_PATH, name, Util::JSON);
 
-		// TODO Dependency check
+		if (Map::checkDependent(actor.string()))
+			return Result::DEPENDENT;
 
 		if (!fs::remove(actor))
 			return Result::DOES_NOT_EXIST;
 		return Result::SUCCESS;
 	}
 
-
-	const Result remove(const std::string& name) {
-		STRING_CHECKS(name);
-		return remove(name.c_str());
-	}
-
 	const std::string list() {
 		return Util::collect(ACTOR_PATH, Util::JSON_FILTER);
+	}
+
+	const Result change(const char* name, const char* key, const char* value) {
+		return change(std::string(name), std::string(key), std::string(value));
+	}
+
+	const Result change(const std::string& name, const std::string& key, const std::string& value) {
+		auto actor = Util::getResource(ACTOR_PATH, name, Util::JSON);
+		if (!fs::exists(actor))
+			return Result::DOES_NOT_EXIST;
+
+		JSON_UPDATE(actor, json[key] = value;);
+		return Result::SUCCESS;
 	}
 }
